Fix faulty test logic
